+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PublisherDashboard/pages/Solutions/NewSolutionFlow/PublishSolutionOutlet.tsx
@@ -53,7 +53,8 @@ const PublishSolutionOutlet = () => {
 
 				if (isContinue) {
 					SOLUTION_FLOW_ITEMS[index].checked = true;
-				} else {
+				}
+				else {
 					SOLUTION_FLOW_ITEMS[index - 1].checked = false;
 				}
